rename: 파일 위치, 이름 변경 - 바텀 시트에서 쓰는 파일들은 전부 bottom-sheet 폴더, 관련 이름으로 변경
