+++ client/app/search/[id]/bottom-sheet/BottomSheet.module.scss
@@ -1,0 +1,16 @@
+.main {
+  display: flex;
+  flex-direction: column;
+  position: fixed;
+  z-index: 1;
+  top: 96%;
+  left: 0;
+  border-top-left-radius: 8px;
+  border-top-right-radius: 8px;
+  background-color: #fff;
+  box-shadow: 0px 0px 10px rgba(0, 0, 0, 0.6);
+  width: 100vw;
+  height: 100%;
+  transition: transform 150ms ease-out;
+  z-index: 9999;
+}
