+++ client/app/search/[id]/bottom-sheet/bottom-sheet-content/GoogleMap.module.scss
