+++ client/app/search/[id]/bottom-sheet/bottom-sheet-content/GoogleMap.tsx
@@ -67,7 +67,7 @@ const GoogleMaps = () => {
                 setSelectedMarker(undefined);
               }}
             >
-              <span>as</span>
+              <span>mansndjs</span>
             </InfoWindowF>
           )}
         </GoogleMap>
