+++ client/app/search/[id]/bottom-sheet/bottom-sheet-header/BottomSheetHeader.module.scss
@@ -1,0 +1,17 @@
+.main {
+  height: 48px;
+  border-top-left-radius: 8px;
+  border-top-right-radius: 8px;
+  position: relative;
+  padding-top: 16px;
+  padding-bottom: 4px;
+}
+
+.handle {
+  width: 50px;
+  height: 4px;
+  border-radius: 2px;
+  background-color: #d0d0d0;
+  margin: auto;
+  margin-bottom: 20px;
+}
