--- client/app/search/[id]/map/map-closer/MapCloser.module.scss
@@ -1,22 +1,0 @@
-.main {
-  position: fixed;
-  top: 0;
-  left: 0;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  width: 100vw;
-  height: 10vh;
-  background-color: rgb(231, 231, 231);
-  padding: 0 1rem;
-  column-gap: 1rem;
-  z-index: 9999;
-}
-
-.close_button {
-  font-size: 26px;
-}
-
-.close_text {
-  font-size: 20px;
-}
